BREAKING CHANGE: moved all security interfaces to subpackage
